Refactor CERM.checkForExistingSpecies using strict=False isomorphism
Since this option allows us to compare resonance structures directly,
we don't need to worry about nonreactive resonance structures or
aromatic resonance structures.

Also, the return value of checkForExistingSpecies is reduced to
a single value, since returning whether or not the species
was found is redundant with returning the species itself.