Refactor some pdep logging statements for improved performance
Instead of pre-formatting strings provided to the logging module,
use sprintf style formatting and provide arguments directly.

This way, the logging module can decide whether or not to
format the strings, which can take a significant amount of time.

Also, this converts a few statements from warning to debug
in order to clean up the log file for pdep jobs.